+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.5.0](https://github.com/act3-ace/run-time-assurance/compare/v1.4.0...v1.5.0) (2023-01-16)
+
+
+### Features
+
+* **implicit asif:** Fix subsample_constraints for implicit ASIF ([4f6a7f4](https://github.com/act3-ace/run-time-assurance/commit/4f6a7f4cc2ddee25c6c9c81c90efd014a47010b3))
+
+# [1.4.0](https://github.com/act3-ace/run-time-assurance/compare/v1.3.0...v1.4.0) (2023-01-16)
+
+
+### Features
+
+* **dynamics:** Default ASIF predicted state method ([e052c7c](https://github.com/act3-ace/run-time-assurance/commit/e052c7c403f8dcbd22a9ffc7d0008b0af2c35339))
+
 # [1.3.0](https://github.com/act3-ace/run-time-assurance/compare/v1.2.0...v1.3.0) (2022-11-14)